make the PGN update process manual
Getting to the CMake setup where the PGNs will be updated in
every case would require quite a bit of work. Since this is something
that won't be done often, reduce the dependencies at build time
(no need for Ruby to build the package) and make the development
process simpler.