complete and debug Receiver and toCAN/fromCAN based on real-world messages
Given how much of a blackbox the Actisense device is, I elected
to go all-in on direct CAN interfacing (which we needed anyway for
the BBB, and was planning to use for the PC)

This commit modifies Receiver to add generic NMEA2000 commands
such as ID allocation and product information queries, and makes
it store device information as it is received.

A bunch of misunderstanding in the initial implementation:
- there is no guarantee that the device will send exactly the number
  of bytes necessary to reassemble a message. The last message can
  be bigger
- the size field in the first message is the number of bytes, not
  the number of messages.

I still don't know if the Actisense device does or does not do
reassembly - reading other people code, it seems that it does not,
but I haven't been able to check. So, for now, do not use Receiver
with Actisense. Once I get my hands on an AIS receiver, I'll know
and may refactor ActisenseDriver to make use of the Receiver facilities
if needed.